int to long fix

+++ SharedCounterApproaches/src/main/java/com/company/Counter.java
@@ -3,6 +3,6 @@ package com.company;
 public interface Counter {
 
     public void increment();
-    public int getValue();
+    public long getValue();
 
 }
